+++ sgl-model-gateway/src/utils/mod.rs
@@ -1,0 +1,4 @@
+pub mod http_utils;
+
+pub use http_utils::AttachedBody;
+
